refactor(421): extract map + set host families; value-type table to collections.py
Step 7 of the codegen/api.py decomposition (#421).  The Map<K,V> and Set<T> families move out of execute() into vera/runtime/map.py and vera/runtime/set.py via register_map/register_set(linker, ops_used).  Both dispatch per element/key/value type through _VAL_WASM_TYPES, which moves to a small shared vera/runtime/collections.py constant module that map.py and set.py import.  The families' nested _define_*/_make_option/_map_put helpers travel inside the register functions; everything else they need (bucket codec, Option/Array/String allocators, _WRAP_KIND_* tags) is already module-level in vera.runtime.heap.  The _host_store_refs-coupled host_decref_handle/host_attach_bucket shared runtime stays inline (it goes with the decimal layer).  api.py 2,171->1,735 lines (-60% from the original 4,358).  mypy + ruff clean; 91 map/set codegen tests pass and ch09_map/ch09_set run end-to-end.
